use smalled types for kanji strokes and radical number

+++ include/kanji_tools/kana/KanaConvert.h
@@ -111,7 +111,7 @@ private:
                                             CharType source, const Set& afterN,
                                             const Set& smallKana) const;
   [[nodiscard]] std::string kanaLetters(const std::string& letterGroup,
-                                        CharType source, size_t count,
+                                        CharType source, u_int8_t count,
                                         const Kana*& prevKana,
                                         bool prolong = false) const;
   [[nodiscard]] std::string convertFromRomaji(const std::string& input) const;
