+++ include/kanji_tools/kanji/Radical.h
@@ -11,7 +11,7 @@ class Radical {
 public:
   using AltForms = std::vector<std::string>;
 
-  Radical(int number, const std::string& name, const AltForms& altForms,
+  Radical(u_int8_t number, const std::string& name, const AltForms& altForms,
           const std::string& longName, const std::string& reading)
       : _number(number), _name(name), _altForms(altForms), _longName(longName),
         _reading(reading) {}
@@ -32,7 +32,7 @@ public:
   [[nodiscard]] auto& longName() const { return _longName; }
   [[nodiscard]] auto& reading() const { return _reading; }
 private:
-  int _number;
+  u_int8_t _number;
   std::string _name;
   AltForms _altForms;
   std::string _longName;
